requirements.txt
telethon==1.28.5
python-dotenv==1.0.0
matplotlib==3.7.1
numpy==1.24.3
python-dateutil==2.8.2
argparse==1.4.0
tabulate==0.9.0